scripts/requirements.txt
openai>=1.0.0
google-generativeai>=0.3.0
Pillow>=9.0.0
requests>=2.25.0
